Move LoadImage method to the LoadImageFunction class.
