+++ MetaphysicsIndustries.Solus.Test/MetaphysicsIndustries.Solus.Test.csproj
@@ -117,6 +117,7 @@
     <Compile Include="FunctionsT\LessThanOrEqualComparisonOperationT\LessThanOrEqualComparisonOperationTest.cs" />
     <Compile Include="FunctionsT\LoadImageFunctionT\GetResultTest.cs" />
     <Compile Include="FunctionsT\LoadImageFunctionT\LoadImageFunctionTest.cs" />
+    <Compile Include="FunctionsT\LoadImageFunctionT\LoadImageTest.cs" />
     <Compile Include="FunctionsT\Log10FunctionT\GetResultTest.cs" />
     <Compile Include="FunctionsT\Log10FunctionT\Log10FunctionTest.cs" />
     <Compile Include="FunctionsT\Log2FunctionT\GetResultTest.cs" />
@@ -158,7 +159,6 @@
     <Compile Include="MockMathObjectF.cs" />
     <Compile Include="MockExpression.cs" />
     <Compile Include="MockMathObject.cs" />
-    <Compile Include="EvaluatorT\LoadImageTest.cs" />
     <Compile Include="SolusParserT\SolusParserTest.cs" />
     <Compile Include="ValuesT\IntervalT\ContainsTest.cs" />
     <Compile Include="ValuesT\IntervalT\EmptyDegenerateTest.cs" />
@@ -187,6 +187,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="EvaluatorT" />
   </ItemGroup>
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
